Move events are now in ShapeSelection component
