+++ index.html
@@ -26,11 +26,8 @@
             height: 800px;
             border: 1px black solid;
         }
-        #selected_elements {
-            pointer-events: none;
-        }
         .selected_rect{
-            fill: none;
+            fill: transparent;
         }
         .draggable {
             cursor: move;
