+++ src/components/FigureList.tsx
@@ -6,7 +6,7 @@ import {AppState} from "../models/AppState";
 
 class FigureList extends React.Component<any, any> {
     render() {
-        return (<svg width="100%" height="100%">
+        return (<svg width="100%" height="100%" id="canvasSvg">
                 <g id="canvas_elements">
                     {this.props.figures.map(
                         (figure: any) =>
